rename nn.loss to nn.metrics
