+++ chemprop/cli/fingerprint.py
@@ -12,7 +12,7 @@ from chemprop.cli.common import add_common_args, process_common_args, validate_c
 from chemprop.cli.predict import find_models
 from chemprop.cli.utils import Subcommand, build_data_from_files, make_dataset
 from chemprop.models import load_model
-from chemprop.nn.loss import LossFunctionRegistry
+from chemprop.nn.metrics import LossFunctionRegistry
 
 logger = logging.getLogger(__name__)
